# openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: My hiit muscle
  license:
    name: MIT
servers:
  - url: http://localhost/v1
paths:
  /muscles:
    get:
      summary: List all muscles
      operationId: listMuscles
      tags:
        - muscles
      parameters:
        - name: limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: A paged array of pets
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/muscles"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a muscle
      operationId: createMuscles
      tags:
        - muscles
      responses:
        '201':
          description: Return muscle created
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /muscles/{muscleId}:
    get:
      summary: Info for a specific muscle
      operationId: showMuscleById
      tags:
        - muscles
      parameters:
        - name: muscleId
          in: path
          required: true
          description: The id of the muscle to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Muscle"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Muscle:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Muscles:
      type: array
      items:
        $ref: "#/components/schemas/Muscle"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string